+++ R/use_ai.R
@@ -27,6 +27,19 @@ use_ai <- function(
   overwrite = FALSE,
   open = rlang::is_interactive(),
   gh_token = gh::gh_token(),
+  allowlist = c(
+    "api.github.com",
+    "api2r.org",
+    "bioconductor.org",
+    "cloud.r-project.org",
+    "CRAN.R-project.org",
+    "docs.github.com",
+    "r-lib.org",
+    "rstudio.github.io",
+    "tidymodels.org",
+    "tidyverse.org",
+    "wrangle.zone"
+  ),
   skills = c(
     "create-issue",
     "document",
@@ -44,7 +57,12 @@ use_ai <- function(
     overwrite = overwrite,
     open = open
   )
-  copilot_path <- use_github_copilot(overwrite = overwrite, open = open)
+  copilot_path <- use_github_copilot(
+    overwrite = overwrite,
+    open = open,
+    allowlist = allowlist,
+    gh_token = gh_token
+  )
 
   skill_fn_names <- stringr::str_c(
     "use_skill_",